fix(sandbox): thread shared canonicalization through multi-root planning
Integration commit for the #3450 issue-family branches.

The #3463 multi-root planner resolved its workspace roots with a private
fs.realpathSync call; route that resolution through the #3475 shared
fail-fast helper so a raced, cyclic, or malformed root fails with the
classified canonicalization error while preserving #3463 mount
correction guidance and propagating the errno cause chain. Add
symlink-cycle include-root behavioral coverage for the merged behavior.

Combining all nine issue branches pushed three shared files past the
eslint size limits that every branch individually respected, so split
along existing seams without behavior change: the wrong-platform
binary-header preflight moves from sandbox-node-modules.ts into
sandbox-binary-preflight.ts; the #3475/#3463 multi-root and
canonicalization describes move to sandbox-node-modules-multiroot.test.ts;
the #3475 mount-source describe moves to
sandbox-containers-mounts.test.ts; and the sandbox image-presence failure
message moves to its own helper in sandbox-exec.ts.
